+++ modules/Users/UserFields.php
@@ -1,378 +1,322 @@
 <?php
-DrawHeader(ProgramTitle());
+/**
+ * User Fields
+ *
+ * @package RosarioSIS
+ * @subpackage modules
+ */
+
+require_once 'ProgramFunctions/Fields.fnc.php';
+
+DrawHeader( ProgramTitle() );
+
 //$_ROSARIO['allow_edit'] = true;
 
-if ( $_REQUEST['tables'] && $_POST['tables'] && AllowEdit())
+if ( isset( $_POST['tables'] )
+	&& is_array( $_POST['tables'] )
+	&& AllowEdit() )
 {
 	$table = $_REQUEST['table'];
-	foreach ( (array) $_REQUEST['tables'] as $id => $columns)
+
+	foreach ( (array) $_REQUEST['tables'] as $id => $columns )
 	{
-//FJ fix SQL bug invalid sort order
-//FJ fix SQL bug invalid display columns
-		if ((empty($columns['SORT_ORDER']) || is_numeric($columns['SORT_ORDER'])) && (empty($columns['COLUMNS']) || is_numeric($columns['COLUMNS'])))
+		// FJ fix SQL bug invalid sort order.
+		if ( ( empty( $columns['SORT_ORDER'] )
+				|| is_numeric( $columns['SORT_ORDER'] ) )
+			&& ( empty( $columns['COLUMNS'] )
+				|| is_numeric( $columns['COLUMNS'] ) ) )
 		{
-			//FJ added SQL constraint TITLE is not null
-			if ((!isset($columns['TITLE']) || !empty($columns['TITLE'])))
+			// FJ added SQL constraint TITLE is not null.
+			if ( ! isset( $columns['TITLE'] )
+				|| ! empty( $columns['TITLE'] ) )
 			{
-				if ( $id!='new')
+				// Update Field / Category.
+				if ( $id !== 'new' )
 				{
-					if ( $columns['CATEGORY_ID'] && $columns['CATEGORY_ID']!=$_REQUEST['category_id'])
+					if ( isset( $columns['CATEGORY_ID'] )
+						&& $columns['CATEGORY_ID'] != $_REQUEST['category_id'] )
+					{
 						$_REQUEST['category_id'] = $columns['CATEGORY_ID'];
+					}
 
-					$sql = "UPDATE $table SET ";
+					$sql = 'UPDATE ' . $table . ' SET ';
 
-					foreach ( (array) $columns as $column => $value)
-						$sql .= $column."='".$value."',";
-					$sql = mb_substr($sql,0,-1) . " WHERE ID='".$id."'";
+					foreach ( (array) $columns as $column => $value )
+					{
+						$sql .= $column . "='" . $value . "',";
+					}
+
+					$sql = mb_substr( $sql, 0, -1 ) . " WHERE ID='" . $id . "'";
+
 					$go = true;
 				}
+				// New Field / Category.
 				else
 				{
-					$sql = "INSERT INTO $table ";
+					$sql = 'INSERT INTO ' . $table . ' ';
 
-					if ( $table=='STAFF_FIELDS')
+					// New Field.
+					if ( $table === 'STAFF_FIELDS' )
 					{
-						if ( $columns['CATEGORY_ID'])
+						if ( isset( $columns['CATEGORY_ID'] ) )
 						{
 							$_REQUEST['category_id'] = $columns['CATEGORY_ID'];
-							unset($columns['CATEGORY_ID']);
+
+							unset( $columns['CATEGORY_ID'] );
 						}
-						$id = DBGet(DBQuery("SELECT ".db_seq_nextval('STAFF_FIELDS_SEQ').' AS ID '.FROM_DUAL));
-						$id = $id[1]['ID'];
-						$fields = "ID,CATEGORY_ID,";
-						$values = $id.",'".$_REQUEST['category_id']."',";
-						$_REQUEST['id'] = $id;
 
-						$create_index = true;
-						switch ( $columns['TYPE'])
-						{
-							case 'radio':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id VARCHAR(1)");
-							break;
+						$_REQUEST['id'] = AddDBField( 'STAFF', 'staff_fields_seq', $columns['TYPE'] );
 
-							case 'text':
-							case 'exports':
-							case 'select':
-							case 'autos':
-							case 'edits':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id VARCHAR(255)");
-							break;
+						$fields = 'ID,CATEGORY_ID,';
 
-							case 'codeds':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id VARCHAR(15)");
-							break;
-
-							case 'multiple':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id VARCHAR(1000)");
-							break;
-
-							case 'numeric':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id NUMERIC(20,2)");
-							break;
-
-							case 'date':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id DATE");
-							break;
-
-							case 'textarea':
-								DBQuery("ALTER TABLE STAFF ADD CUSTOM_$id VARCHAR(5000)");
-								$create_index = false; //FJ SQL bugfix index row size exceeds maximum 2712 for index
-							break;
-						}
-						if ( $create_index)
-							DBQuery("CREATE INDEX STAFF_IND$id ON STAFF (CUSTOM_$id)");
+						$values = $_REQUEST['id'] . ",'" . $_REQUEST['category_id'] . "',";
 					}
-					elseif ( $table=='STAFF_FIELD_CATEGORIES')
+					// New Category.
+					elseif ( $table === 'STAFF_FIELD_CATEGORIES' )
 					{
-						$id = DBGet(DBQuery("SELECT ".db_seq_nextval('STAFF_FIELD_CATEGORIES_SEQ').' AS ID '.FROM_DUAL));
+						$id = DBGet( DBQuery( 'SELECT ' . db_seq_nextval( 'STAFF_FIELD_CATEGORIES_SEQ' ) . ' AS ID ' . FROM_DUAL ) );
+
 						$id = $id[1]['ID'];
+
 						$fields = "ID,";
-						$values = $id.",";
+
+						$values = $id . ",";
+
 						$_REQUEST['category_id'] = $id;
-						// add to profile or permissions of user creating it
-						if (User('PROFILE_ID'))
-							DBQuery("INSERT INTO PROFILE_EXCEPTIONS (PROFILE_ID,MODNAME,CAN_USE,CAN_EDIT) values('".User('PROFILE_ID')."','Users/User.php&category_id=$id','Y','Y')");
+
+						// Add to profile or permissions of user creating it.
+						if ( User( 'PROFILE_ID' ) )
+						{
+							DBQuery( "INSERT INTO PROFILE_EXCEPTIONS (PROFILE_ID,MODNAME,CAN_USE,CAN_EDIT)
+								values('" . User( 'PROFILE_ID' ) . "','Users/User.php&category_id=" . $id . "','Y','Y')" );
+						}
 						else
-							DBQuery("INSERT INTO STAFF_EXCEPTIONS (USER_ID,MODNAME,CAN_USE,CAN_EDIT) values('".User('STAFF_ID')."','Users/User.php&category_id=$id','Y','Y')");
+						{
+							DBQuery( "INSERT INTO STAFF_EXCEPTIONS (USER_ID,MODNAME,CAN_USE,CAN_EDIT)
+								values('" . User( 'STAFF_ID' ) . "','Users/User.php&category_id=" . $id . "','Y','Y')" );
+						}
 					}
 
 					$go = false;
 
-					foreach ( (array) $columns as $column => $value)
+					foreach ( (array) $columns as $column => $value )
 					{
-						if ( !empty($value) || $value=='0')
+						if ( ! empty( $value )
+							|| $value == '0' )
 						{
-							$fields .= $column.',';
-							$values .= "'".$value."',";
+							$fields .= $column . ',';
+
+							$values .= "'" . $value . "',";
+
 							$go = true;
 						}
 					}
-					$sql .= '(' . mb_substr($fields,0,-1) . ') values(' . mb_substr($values,0,-1) . ')';
+					$sql .= '(' . mb_substr( $fields, 0, -1 ) . ') values(' . mb_substr( $values, 0, -1 ) . ')';
 				}
 
-				if ( $go)
-					DBQuery($sql);
+				if ( $go )
+				{
+					DBQuery( $sql );
+				}
 			}
 			else
-				$error[] = _('Please fill in the required fields');
+				$error[] = _( 'Please fill in the required fields' );
 		}
 		else
-			$error[] = _('Please enter valid Numeric data.');
+			$error[] = _( 'Please enter valid Numeric data.' );
 	}
-	unset($_REQUEST['tables']);
+
+	unset( $_REQUEST['tables'] );
 }
 
-if ( $_REQUEST['modfunc']=='delete' && AllowEdit())
+// Delete Field / Category.
+if ( $_REQUEST['modfunc'] == 'delete'
+	&& AllowEdit() )
 {
-	if ( $_REQUEST['id'])
+	if ( isset( $_REQUEST['id'] )
+		&& intval( $_REQUEST['id'] ) > 0 )
 	{
-		if (DeletePrompt(_('User Field')))
+		if ( DeletePrompt( _( 'User Field' ) ) )
 		{
-			$id = $_REQUEST['id'];
-			DBQuery("DELETE FROM STAFF_FIELDS WHERE ID='".$id."'");
-			DBQuery("ALTER TABLE STAFF DROP COLUMN CUSTOM_$id");
-			$_REQUEST['modfunc'] = '';
-			unset($_REQUEST['id']);
+			DeleteDBField( 'STAFF', $_REQUEST['id'] );
+
+			$_REQUEST['modfunc'] = false;
+
+			unset( $_REQUEST['id'] );
 		}
 	}
-	elseif ( $_REQUEST['category_id'])
+	elseif ( isset( $_REQUEST['category_id'] )
+		&& intval( $_REQUEST['category_id'] ) > 0 )
 	{
-		if (DeletePrompt(_('User Field Category').' '._('and all fields in the category')))
+		if ( DeletePrompt( _( 'User Field Category' ) . ' ' .
+				_( 'and all fields in the category' ) ) )
 		{
-			$fields = DBGet(DBQuery("SELECT ID FROM STAFF_FIELDS WHERE CATEGORY_ID='".$_REQUEST['category_id']."'"));
-			foreach ( (array) $fields as $field)
-			{
-				DBQuery("DELETE FROM STAFF_FIELDS WHERE ID='".$field['ID']."'");
-				DBQuery("ALTER TABLE STAFF DROP COLUMN CUSTOM_$field[ID]");
-			}
-			DBQuery("DELETE FROM STAFF_FIELD_CATEGORIES WHERE ID='".$_REQUEST['category_id']."'");
-			// remove from profiles and permissions
-			DBQuery("DELETE FROM PROFILE_EXCEPTIONS WHERE MODNAME='Users/User.php&category_id=$_REQUEST[category_id]'");
-			DBQuery("DELETE FROM STAFF_EXCEPTIONS WHERE MODNAME='Users/User.php&category_id=$_REQUEST[category_id]'");
-			$_REQUEST['modfunc'] = '';
-			unset($_REQUEST['category_id']);
+			DeleteDBFieldCategory( 'STAFF', $_REQUEST['category_id'] );
+
+			// Remove from profiles and permissions.
+			DBQuery( "DELETE FROM PROFILE_EXCEPTIONS
+				WHERE MODNAME='Users/User.php&category_id=" . $_REQUEST['category_id'] . "'" );
+
+			DBQuery( "DELETE FROM STAFF_EXCEPTIONS
+				WHERE MODNAME='Users/User.php&category_id=" . $_REQUEST['category_id'] . "'" );
+
+			$_REQUEST['modfunc'] = false;
+
+			unset( $_REQUEST['category_id'] );
 		}
 	}
 }
 
-if (empty($_REQUEST['modfunc']))
-
+if ( ! $_REQUEST['modfunc'] )
 {
-//FJ fix SQL bug invalid sort order
-	if (isset($error)) 
-		echo ErrorMessage($error);
-		
-	// CATEGORIES
-	$sql = "SELECT ID,TITLE,SORT_ORDER FROM STAFF_FIELD_CATEGORIES ORDER BY SORT_ORDER,TITLE";
-	$QI = DBQuery($sql);
-	$categories_RET = DBGet($QI);
-
-	if (AllowEdit() && $_REQUEST['id']!='new' && $_REQUEST['category_id']!='new' && ($_REQUEST['id'] || $_REQUEST['category_id']>2))
+	if ( isset( $error ) ) 
 	{
-		$delete_URL = "'Modules.php?modname=" . $_REQUEST['modname'] .
-			'&modfunc=delete&category_id=' . $_REQUEST['category_id'] .
-			'&id=' . $_REQUEST['id'] . "'";
-
-		$delete_button = '<input type="button" value="' . _( 'Delete' ) . '" onClick="javascript:ajaxLink(' . $delete_URL . ');" />';
+		echo ErrorMessage( $error );
 	}
-
-	// ADDING & EDITING FORM
-	if ( $_REQUEST['id'] && $_REQUEST['id']!='new')
+	
+	// ADDING & EDITING FORM.
+	if ( $_REQUEST['id']
+		&& $_REQUEST['id'] !== 'new' )
 	{
-		$sql = "SELECT CATEGORY_ID,TITLE,TYPE,SELECT_OPTIONS,DEFAULT_SELECTION,SORT_ORDER,REQUIRED,REQUIRED,(SELECT TITLE FROM STAFF_FIELD_CATEGORIES WHERE ID=CATEGORY_ID) AS CATEGORY_TITLE FROM STAFF_FIELDS WHERE ID='".$_REQUEST['id']."'";
-		$RET = DBGet(DBQuery($sql));
-		$RET = $RET[1];
-		$title = ParseMLField($RET['CATEGORY_TITLE']).' - '.ParseMLField($RET['TITLE']);
-	}
-	elseif ( $_REQUEST['category_id'] && $_REQUEST['category_id']!='new' && $_REQUEST['id']!='new')
-	{
-		$sql = "SELECT TITLE,ADMIN,TEACHER,PARENT,NONE,SORT_ORDER,INCLUDE,COLUMNS
+		$RET = DBGet( DBQuery( "SELECT ID,CATEGORY_ID,TITLE,TYPE,SELECT_OPTIONS,
+			DEFAULT_SELECTION,SORT_ORDER,REQUIRED,
+			(SELECT TITLE
 				FROM STAFF_FIELD_CATEGORIES
-				WHERE ID='".$_REQUEST['category_id']."'";
-		$RET = DBGet(DBQuery($sql));
+				WHERE ID=CATEGORY_ID) AS CATEGORY_TITLE
+			FROM STAFF_FIELDS
+			WHERE ID='" . $_REQUEST['id'] . "'" ) );
+
 		$RET = $RET[1];
-		$title = ParseMLField($RET['TITLE']);
-	}
-	elseif ( $_REQUEST['id']=='new')
-		$title = _('New User Field');
-	elseif ( $_REQUEST['category_id']=='new')
-		$title = _('New User Field Category');
 
-	if ( $_REQUEST['id'])
+		$title = ParseMLField( $RET['CATEGORY_TITLE'] ) . ' - ' . ParseMLField( $RET['TITLE'] );
+	}
+	elseif ( $_REQUEST['category_id']
+		&& $_REQUEST['category_id'] !== 'new'
+		&& $_REQUEST['id'] !== 'new' )
 	{
-		echo '<form action="Modules.php?modname='.$_REQUEST['modname'].'&category_id='.$_REQUEST['category_id'];
+		$RET = DBGet( DBQuery( "SELECT ID AS CATEGORY_ID,TITLE,ADMIN,TEACHER,PARENT,NONE,SORT_ORDER,INCLUDE,COLUMNS
+			FROM STAFF_FIELD_CATEGORIES
+			WHERE ID='" . $_REQUEST['category_id'] . "'" ) );
 
-		if ( $_REQUEST['id']!='new')
-			echo '&id='.$_REQUEST['id'];
+		$RET = $RET[1];
 
-		echo '&table=STAFF_FIELDS" method="POST">';
-
-		DrawHeader($title,$delete_button.SubmitButton(_('Save')));
-
-		$header .= '<table class="width-100p valign-top fixed-col"><tr class="st">';
-
-		//FJ field name required
-		$header .= '<td>' . MLTextInput($RET['TITLE'],'tables['.$_REQUEST['id'].'][TITLE]',(! $RET['TITLE']?'<span class="legend-red">':'')._('Field Name').(! $RET['TITLE']?'</span>':'')) . '</td>';
-
-		// You can't change a user field type after it has been created
-		// mab - allow changing between select and autos and edits and text and exports
-		if ( $_REQUEST['id']!='new')
-		{
-			if ( $RET['TYPE']!='select' && $RET['TYPE']!='autos' && $RET['TYPE']!='edits' && $RET['TYPE']!='text' && $RET['TYPE']!='exports')
-			{
-				$allow_edit = $_ROSARIO['allow_edit'];
-				$AllowEdit = $_ROSARIO['AllowEdit'][ $modname ];
-				$_ROSARIO['allow_edit'] = false;
-				$_ROSARIO['AllowEdit'][ $modname ] = array();
-				$type_options = array('select' => _('Pull-Down'),'autos' => _('Auto Pull-Down'),'edits' => _('Edit Pull-Down'),'text' => _('Text'),'radio' => _('Checkbox'),'codeds' => _('Coded Pull-Down'),'exports' => _('Export Pull-Down'),'numeric' => _('Number'),'multiple' => _('Select Multiple from Options'),'date' => _('Date'),'textarea' => _('Long Text'));
-			}
-			else
-				$type_options = array('select' => _('Pull-Down'),'autos' => _('Auto Pull-Down'),'edits' => _('Edit Pull-Down'),'exports' => _('Export Pull-Down'),'text' => _('Text'));
-		}
-		else
-			$type_options = array('select' => _('Pull-Down'),'autos' => _('Auto Pull-Down'),'edits' => _('Edit Pull-Down'),'text' => _('Text'),'radio' => _('Checkbox'),'codeds' => _('Coded Pull-Down'),'exports' => _('Export Pull-Down'),'numeric' => _('Number'),'multiple' => _('Select Multiple from Options'),'date' => _('Date'),'textarea' => _('Long Text'));
-
-		$header .= '<td>' . SelectInput($RET['TYPE'],'tables['.$_REQUEST['id'].'][TYPE]',_('Data Type'),$type_options,false) . '</td>';
-		if ( $_REQUEST['id']!='new' && $RET['TYPE']!='select' && $RET['TYPE']!='autos' && $RET['TYPE']!='edits' && $RET['TYPE']!='text' && $RET['TYPE']!='exports')
-		{
-			$_ROSARIO['allow_edit'] = $allow_edit;
-			$_ROSARIO['AllowEdit'][ $modname ] = $AllowEdit;
-		}
-		foreach ( (array) $categories_RET as $type)
-			$categories_options[$type['ID']] = $type['TITLE'];
-
-		$header .= '<td>' . MLSelectInput($RET['CATEGORY_ID']?$RET['CATEGORY_ID']:$_REQUEST['category_id'],'tables['.$_REQUEST['id'].'][CATEGORY_ID]',_('User Field Category'),$categories_options,false) . '</td>';
-
-		$header .= '<td>' . TextInput($RET['SORT_ORDER'],'tables['.$_REQUEST['id'].'][SORT_ORDER]',_('Sort Order'),'size=5') . '</td>';
-
-		$header .= '</tr><tr class="st">';
-		$colspan = 2;
-		if ( $RET['TYPE']=='autos' || $RET['TYPE']=='edits' || $RET['TYPE']=='select' || $RET['TYPE']=='codeds' || $RET['TYPE']=='multiple' || $RET['TYPE']=='exports' || $_REQUEST['id']=='new')
-		{
-			$header .= '<td colspan="2">'.TextAreaInput($RET['SELECT_OPTIONS'],'tables['.$_REQUEST['id'].'][SELECT_OPTIONS]',_('Pull-Down').'/'._('Auto Pull-Down').'/'._('Coded Pull-Down').'/'._('Select Multiple from Options').'<br />'._('* one per line'),'rows=7 cols=40') . '</td>';
-			$colspan = 1;
-		}
-		$header .= '<td style="vertical-align:bottom;" colspan="'.$colspan.'">'.TextInput($RET['DEFAULT_SELECTION'],'tables['.$_REQUEST['id'].'][DEFAULT_SELECTION]',_('Default')).'<br />'._('* for dates: YYYY-MM-DD').',<br />&nbsp;'._('for checkboxes: Y').'</td>';
-
-		$new = ($_REQUEST['id']=='new');
-		$header .= '<td>' . CheckboxInput($RET['REQUIRED'],'tables['.$_REQUEST['id'].'][REQUIRED]',_('Required'),'',$new) . '</td>';
-
-		$header .= '</tr></table>';
+		$title = ParseMLField( $RET['TITLE'] );
 	}
-	elseif ( $_REQUEST['category_id'])
+	elseif ( $_REQUEST['id'] === 'new' )
 	{
-		echo '<form action="Modules.php?modname='.$_REQUEST['modname'].'&table=STAFF_FIELD_CATEGORIES';
+		$title = _( 'New User Field' );
 
-		if ( $_REQUEST['category_id']!='new')
-			echo '&category_id='.$_REQUEST['category_id'];
+		$RET['ID'] = 'new';
 
-		echo '" method="POST">';
+		$RET['CATEGORY_ID'] = $_REQUEST['category_id'];
+	}
+	elseif ( $_REQUEST['category_id'] === 'new' )
+	{
+		$title = _( 'New User Field Category' );
 
-		DrawHeader($title,$delete_button.SubmitButton(_('Save')));
+		$RET['CATEGORY_ID'] = 'new';
+	}
 
-		$header .= '<table class="width-100p valign-top"><tr class="st">';
+	if ( $_REQUEST['category_id']
+		&& ! $_REQUEST['id'] )
+	{
+		$extra_fields = array();
 
-		//FJ title required
-		$header .= '<td>' . MLTextInput($RET['TITLE'],'tables['.$_REQUEST['category_id'].'][TITLE]',(! $RET['TITLE']?'<span class="legend-red">':'')._('Title').(! $RET['TITLE']?'</span>':'')) . '</td>';
+		$extra_fields[] = TextInput(
+			$RET['COLUMNS'],
+			'tables[' . $category_id . '][COLUMNS]',
+			_( 'Display Columns' ),
+			'size=5'
+		);
 
-		$header .= '<td>' . TextInput($RET['SORT_ORDER'],'tables['.$_REQUEST['category_id'].'][SORT_ORDER]',_('Sort Order'),'size=5') . '</td>';
-
-		$header .= '<td>' . TextInput($RET['COLUMNS'],'tables['.$_REQUEST['category_id'].'][COLUMNS]',_('Display Columns'),'size=5') . '</td>';
-
-		$new = ($_REQUEST['category_id']=='new');
-
-		$header .= '<td><table><tr>';
-
-		$header .= '<td>' . CheckboxInput($RET['ADMIN'], 'tables['.$_REQUEST['category_id'].'][ADMIN]', ($_REQUEST['category_id']=='1'&&! $RET['ADMIN']?'<span class="legend-red">':'')._('Administrator').($_REQUEST['category_id']=='1'&&! $RET['ADMIN']?'</span>':''), '', $new, button('check'), button('x')) . '</td>';
-
-		$header .= '<td>' . CheckboxInput($RET['TEACHER'], 'tables['.$_REQUEST['category_id'].'][TEACHER]', ($_REQUEST['category_id']=='1'&&! $RET['TEACHER']?'<span class="legend-red">':'')._('Teacher').($_REQUEST['category_id']=='1'&&! $RET['TEACHER']?'</span>':''), '', $new, button('check'), button('x')) . '</td>';
-
-		$header .= '<td>' . CheckboxInput($RET['PARENT'], 'tables['.$_REQUEST['category_id'].'][PARENT]', ($_REQUEST['category_id']=='1'&&! $RET['PARENT']?'<span class="legend-red">':'')._('Parent').($_REQUEST['category_id']=='1'&&! $RET['PARENT']?'</span>':''), '', $new, button('check'), button('x')) . '</td>';
-
-		$header .= '<td>' . CheckboxInput($RET['NONE'], 'tables['.$_REQUEST['category_id'].'][NONE]', ($_REQUEST['category_id']=='1'&&! $RET['NONE']?'<span class="legend-red">':'')._('No Access').($_REQUEST['category_id']=='1'&&! $RET['NONE']?'</span>':''), '', $new, button('check'), button('x')) . '</td>';
-
-		$header .= '</tr>';
-		$header .= '<tr><td colspan="4"><span class="legend-gray">'._('Profiles').'</span></td></tr>';
-		$header .= '</table></td>';
-
-		if ( $_REQUEST['category_id']>2 || $new)
+		if ( $_REQUEST['category_id'] != 1 )
 		{
-			$header .= '</tr><tr>';
-			$header .= '<td colspan="2"></td>';
-			$header .= '<td>' . TextInput($RET['INCLUDE'],'tables['.$_REQUEST['category_id'].'][INCLUDE]',_('Include (should be left blank for most categories)')) . '</td>';
+			$extra_fields[] = '<table><tr class="st"><td>' .
+				CheckboxInput(
+					$RET['ADMIN'],
+					'tables[' . $_REQUEST['category_id'] . '][ADMIN]',
+					_( 'Administrator' ),
+					'',
+					$_REQUEST['category_id'] === 'new',
+					button( 'check' ),
+					button( 'x' )
+				) . '</td><td>' .
+				CheckboxInput(
+					$RET['TEACHER'],
+					'tables[' . $_REQUEST['category_id'] . '][TEACHER]',
+					_( 'Teacher' ),
+					'',
+					$_REQUEST['category_id'] === 'new',
+					button( 'check' ),
+					button( 'x' )
+				) . '</td></tr><tr><td>' .
+				CheckboxInput(
+					$RET['PARENT'],
+					'tables[' . $_REQUEST['category_id'] . '][PARENT]',
+					_( 'Parent' ),
+					'',
+					$_REQUEST['category_id'] === 'new',
+					button( 'check' ),
+					button( 'x' )
+				) . '</td><td>' .
+				CheckboxInput(
+					$RET['NONE'],
+					'tables[' . $_REQUEST['category_id'] . '][NONE]',
+					_( 'No Access' ),
+					'',
+					$_REQUEST['category_id'] === 'new',
+					button( 'check' ),
+					button( 'x' )
+				) . '</td></tr></table>' .
+				str_replace( '<br />', '', FormatInputTitle(
+					_( 'Profiles' )
+				) );
 		}
 
-		$header .= '</tr></table>';
-	}
-	else
-		$header = false;
-
-	if ( $header)
-	{
-		DrawHeader($header);
-		echo '</form>';
-	}
-
-	// DISPLAY THE MENU
-	$LO_options = array('save'=>false,'search'=>false,'responsive'=>false);
-
-	if (count($categories_RET))
-	{
-		if ( $_REQUEST['category_id'])
+		if ( $_REQUEST['category_id'] > 2
+			|| $_REQUEST['category_id'] === 'new' )
 		{
-			foreach ( (array) $categories_RET as $key => $value)
-			{
-				if ( $value['ID']==$_REQUEST['category_id'])
-					$categories_RET[ $key ]['row_color'] = Preferences('HIGHLIGHT');
-			}
+			$extra_fields[] = TextInput(
+				$RET['INCLUDE'],
+				'tables[' . $category_id . '][INCLUDE]',
+				_( 'Include (should be left blank for most categories)' )
+			);
 		}
 	}
 
+	echo GetFieldsForm(
+		'STAFF',
+		$title,
+		$RET,
+		isset( $extra_fields ) ? $extra_fields : array()
+	);
+
+	// CATEGORIES.
+	$categories_RET = DBGet( DBQuery( "SELECT ID,TITLE,SORT_ORDER
+		FROM STAFF_FIELD_CATEGORIES
+		ORDER BY SORT_ORDER,TITLE" ) );
+
+	// DISPLAY THE MENU.
 	echo '<div class="st">';
-	$columns = array('TITLE' => _('Category'),'SORT_ORDER' => _('Order'));
-	$link = array();
-	$link['TITLE']['link'] = 'Modules.php?modname='.$_REQUEST['modname'].'&modfunc='.$_REQUEST['modfunc'];
-	$link['TITLE']['variables'] = array('category_id' => 'ID');
-	$link['add']['link'] = 'Modules.php?modname='.$_REQUEST['modname'].'&category_id=new';
 
-    $categories_RET = ParseMLArray($categories_RET,'TITLE');
-	ListOutput($categories_RET,$columns,'User Field Category','User Field Categories',$link,array(),$LO_options);
+	FieldsMenuOutput( $categories_RET, $_REQUEST['category_id'] );
+
 	echo '</div>';
 
-	// FIELDS
-	if ( $_REQUEST['category_id'] && $_REQUEST['category_id']!='new' && count($categories_RET))
+	// FIELDS.
+	if ( $_REQUEST['category_id']
+		&& $_REQUEST['category_id'] !=='new'
+		&& $categories_RET )
 	{
-		$sql = "SELECT ID,TITLE,TYPE,SORT_ORDER FROM STAFF_FIELDS WHERE CATEGORY_ID='".$_REQUEST['category_id']."' ORDER BY SORT_ORDER,TITLE";
-		$fields_RET = DBGet(DBQuery($sql),array('TYPE' => '_makeType'));
-
-		if (count($fields_RET))
-		{
-			if ( $_REQUEST['id'] && $_REQUEST['id']!='new')
-			{
-				foreach ( (array) $fields_RET as $key => $value)
-				{
-					if ( $value['ID']==$_REQUEST['id'])
-						$fields_RET[ $key ]['row_color'] = Preferences('HIGHLIGHT');
-				}
-			}
-		}
+		$fields_RET = DBGet( DBQuery( "SELECT ID,TITLE,TYPE,SORT_ORDER
+			FROM STAFF_FIELDS
+			WHERE CATEGORY_ID='" . $_REQUEST['category_id'] . "'
+			ORDER BY SORT_ORDER,TITLE" ), array( 'TYPE' => 'MakeFieldType' ) );
 
 		echo '<div class="st">';
-		$columns = array('TITLE' => _('User Field'),'SORT_ORDER' => _('Order'),'TYPE' => _('Data Type'));
-		$link = array();
-		$link['TITLE']['link'] = 'Modules.php?modname='.$_REQUEST['modname'].'&category_id='.$_REQUEST['category_id'];
-		$link['TITLE']['variables'] = array('id' => 'ID');
-		$link['add']['link'] = 'Modules.php?modname='.$_REQUEST['modname'].'&category_id='.$_REQUEST['category_id'].'&id=new';
 
-        $fields_RET = ParseMLArray($fields_RET,'TITLE');
-		ListOutput($fields_RET,$columns,'User Field','User Fields',$link,array(),$LO_options);
+		FieldsMenuOutput( $fields_RET, $_REQUEST['id'], $_REQUEST['category_id'] );
 
 		echo '</div>';
 	}
-}
-
-function _makeType($value,$name)
-{
-	$options = array('radio' => _('Checkbox'),'text' => _('Text'),'autos' => _('Auto Pull-Down'),'edits' => _('Edit Pull-Down'),'select' => _('Pull-Down'),'codeds' => _('Coded Pull-Down'),'exports' => _('Export Pull-Down'),'date' => _('Date'),'numeric' => _('Number'),'textarea' => _('Long Text'),'multiple' => _('Select Multiple'));
-	return $options[ $value ];
 }