+++ CHANGES.md
@@ -51,6 +51,7 @@ Changes in 2.9
 - Add Photo on mouse over Student Name in InputFinalGrades.php & Grades.php
 - Highlight color: add CSS to HTML head in Warehouse.php
 - Create school_fields_seq Sequence in rosariosis.sql & Update.fnc.php
+- Add Fields (and Field Categories) functions in ProgramFunctions/Fields.fnc.php
 
 Changes in 2.8.11
 -----------------